+++ packages/expo-router/src/useScreens.tsx
@@ -55,7 +55,6 @@ export type ScreenProps<
 > = {
   /** Name is required when used inside a Layout component. */
   name?: string;
-  initialParams?: Record<string, any>;
   options?:
     | TOptions
     | ((prop: { route: DescriptorRouteProp<ParamListBase, string>; navigation: any }) => TOptions);
@@ -97,7 +96,7 @@ function getSortedChildren<
   const entries = [...children];
 
   const ordered = order
-    .map(({ name, initialParams, listeners, options, getId, dangerouslySingular: singular }) => {
+    .map(({ name, listeners, options, getId, dangerouslySingular: singular }) => {
       if (!entries.length) {
         console.warn(`[Layout children]: Too many screens defined. Route "${name}" is extraneous.`);
         return null;
@@ -136,7 +135,7 @@ function getSortedChildren<
 
         return {
           route: match,
-          props: { initialParams, listeners, options, getId },
+          props: { listeners, options, getId },
           routeSource: 'layout' as const,
         };
       }